Hoteldesc/naeem

+++ server/controllers/posts.js
@@ -14,7 +14,10 @@ const createPost = async (req, res) => {
 
 const getAllPosts = async (req, res) => {
   try {
-    const posts = await Post.find().populate("userId", "username email");
+    const posts = await Post.find().populate(
+      "userId",
+      "username email avatar firstName lastName"
+    );
     res.json(posts);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -38,14 +41,21 @@ const getPostById = async (req, res) => {
 const updatePost = async (req, res) => {
   const { id } = req.params;
   try {
+    const post = await Post.findById(id);
+    if (!post) {
+      return res.status(404).json({ message: `Post with id ${id} not found.` });
+    }
+
+    if (!req.user || post.userId.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "User is not authorized to update this post." });
+    }
+
     const updatedPost = await Post.findByIdAndUpdate(id, req.body, {
       new: true,
     });
-    if (!updatedPost) {
-      res.status(404).json({ message: `Post with id ${id} Not Found` });
-    } else {
-      res.json(updatedPost);
-    }
+    res.json(updatedPost);
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -54,12 +64,22 @@ const updatePost = async (req, res) => {
 const deletePost = async (req, res) => {
   const { id } = req.params;
   try {
-    const deletedPost = await Post.findByIdAndDelete(id);
-    if (!deletedPost) {
-      res.status(404).json({ message: `Post with id ${id} Not Found` });
-    } else {
-      res.json(deletedPost);
+    // Fetch the post to check ownership
+    const post = await Post.findById(id);
+    if (!post) {
+      return res.status(404).json({ message: `Post with id ${id} not found.` });
     }
+
+    // Check if the logged-in user is the owner of the post
+    if (!req.user || post.userId.toString() !== req.user._id.toString()) {
+      return res
+        .status(403)
+        .json({ message: "User is not authorized to delete this post." });
+    }
+
+    // If the user is the owner, proceed with the delete
+    const deletedPost = await Post.findByIdAndDelete(id);
+    res.json({ message: "Post successfully deleted", post: deletedPost });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
